requirements/prod.txt
Django==5.2.7
requests==2.32.5
python-dotenv==1.1.1